CHROMIUM: egl/dri2: add support for image, config query, and robustness extensions
This patch enables exposing image, config query and robustness
extensions if swrast driver is used, to improve feature completeness of
software rendering and make it usable by Freon running in a virtual
machine. In addition, swrast is now used as a fallback for
platform_surfaceless if no other driver can be used.

BUG=chromium:394868
TEST=None

Signed-off-by: Prince Agyeman <prince.agyeman@intel.com>
Signed-off-by: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>
Signed-off-by: James Ausmus <james.ausmus@intel.com>
Signed-off-by: Tomasz Figa <tfiga@chromium.org>