HACK: egl/android: Handle HAL_PIXEL_FORMAT_IMPLEMENTATION_DEFINED as RGBX_8888
We don't have proper API to query such buffers for the real internal
format in a generic, scalable way, so let's just hardcode this format to
RGBX_8888 for the time being, in the same way as we hardcoded it in
drm_gralloc.

TODO: Revert when we have the necessary API in place (see b/29886289).

BUG=b:28671744

Change-Id: I492e6045663ea2faa3b5b93821bfb78d7bfe5c25
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/357733
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>