UPSTREAM: egl: Combine the GL and GLES current contexts (v2)
Only keep track of a single current context, instead of separate
contexts for GL and GLES.

In EGL 1.4 (and 1.5), EGL_OPENGL_API and EGL_OPENGL_ES_API are supposed
to be interchangeable for all purposes except for eglCreateContext.

The _EGLThreadInfo::CurrentContexts array is now a single pointer to the
current context, which may be a GL or GLES context. In addition, it now
keeps track of the current API as an enum instead of an index.

eglMakeCurrent will now replace the current context, regardless of which
client API is used for for the current and new contexts. It no longer
checks for a conflicting context. In addition, calling eglMakeCurrent
with EGL_NO_CONTEXT will now release the current context regardless of
the current API.

BUG=b/33495228
TEST=dEQP-EGL.functiona.negative_api.bind_api

v2: Rebased against master (Adam Jackson)

Reviewed-by: Adam Jackson <ajax@redhat.com>
[chromium: Cherry pick to arc-12.1.0-pre2].
(cherry picked from commit 6e066f76ee22909b0653ff8b89429de819e60f14)
Change-Id: I4c7593e279159a7d0f8bc7de1fb53262587721ae
Depends-On: Iba127c3cb0982bdcb08d5ab72f2f0dbf8fd35a25
Reviewed-on: https://chromium-review.googlesource.com/419576
Commit-Ready: Chad Versace <chadversary@chromium.org>
Tested-by: Chad Versace <chadversary@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>