HACK: egl/android: Handle HAL_PIXEL_FORMAT_YCbCr_420_888 as YV12
We don't have proper API to query such buffers for the real internal
format in a generic, scalable way, so let's just hardcode this format to
YV12 for the time being, in the same way as we hardcoded it in drm_gralloc
for all platforms using Mesa.

TODO: Revert when we have the necessary API in place (see b/29886289).

BUG=b:29061386
TEST=CTS:MediaCodecCapabilitiesTest#testAllNonTunneledVideoCodecsSupportFlexibleYUV

Change-Id: If19199c695335d6c81fbf1839f4c8f7fc2b32cc1
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/362736
Reviewed-by: Owen Lin <owenlin@chromium.org>