CHROMIUM: egl/android: Add fallback to kms_swrast driver
If no hardware driver is present, it is possible to fall back to
the kms_swrast driver with any DRI node that supports dumb GEM create
and mmap IOCTLs with softpipe/llvmpipe drivers. This patch makes the
Android EGL platform code retry probe with kms_swrast if hardware-only
probe fails.

BUG=b:28803368
TEST=Boot cyan-cheets in qemu and see the container working

Change-Id: Ida864d90a715d563090c0eb5caa8557684a8cb26
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/374309
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>