FROMLIST: egl/android: Use ANativeWindow::cancelBuffer in destroySurface
When destroying a surface, we might have a buffer dequeued (and possibly used
for rendering) that was not posted on the screen by swapBuffers. We cannot
simply queue the buffer as usual, because it might (and actually it does)
confuse the consumer, who has not requested a swap. For this purpose,
ANativeWindow has a method called cancelBuffer, which we should use instead.

Signed-off-by: Tomasz Figa <tfiga@chromium.org>
(am from https://patchwork.freedesktop.org/patch/151163/)

BUG=b:37406655
TEST=./cts-tradefed run cts --skip-device-info --skip-preconditions -m CtsMediaTest -t android.media.cts.EncodeDecodeTest

Change-Id: I4145874261e6d7ce489b30b1fb37948572d0e42f
Reviewed-on: https://chromium-review.googlesource.com/481559
Commit-Ready: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>