st/dri: Add fence extension to SW path
Use the same fence implementation for drisw.c as dri2.c by making
dri2FenceExtension an external variable. Since the fence implementation
is not dri2.c specific, put it in a separate file. This is desirable for
synchronization in virtual machines.

v2: Don't depend on dri2.c for extensions (Emil)

Change-Id: Ifeaa3c41756f81bd56f9893cd6145c7d014b3074