Use existing calls for getting plane format modifiers.
Use the mesa calls which can retrieve the format modifiers
for a bo instead of a stubbed invalid value of 0.

We no longer need to stub these as mesa has a similar
function.

In d64 this caused issues creating framebuffers as chrome
was passing in 0 which was incompatible with what was created.

See OVER-6083

Change-Id: Ie3dac570c4b26d5e6a27c44d7af6a7d5e044e9dc