Make plane modifiers difference check non-fatal.
I have seen users only consider the first modifier when importing
the dmabuf. Instead of failing just warn and use the first modifier.
If users end up having issues the logs can help determine
the inconsistencies.
This was not made in response to an actual use case, it is
done in anticipation of the potential issue.

OVER-6267