neverware: fix BO import FD conditional
The conditional here was a little weird, maybe it was supposed to be

    fd_data->fds[i] == 0 || fd_data->fds[i] == -1

I think that as written the loop didn't terminate early enough, which
was causing Xwayland windows to show up completely black.

An FD of zero is also valid I think, so simplified the conditional to
just check for -1. The browser side already sets unused entries in the
array to -1, so no change is needed there.