neverware: hacky fix for graphics corruption on iMac7,1
I'm not particularly convinced that this pass is the "real" problem;
by perturbing the shader I can get the bug to reproduce even with this
pass disabled. However, for the non-artificial test cases where I've
been able to reproduce the bug, removing this optimization pass does
seem to fix the problem.

Much more detail in the bug report:

https://neverware.atlassian.net/browse/OVER-7164

OVER-7164