neverware: ignore nouveau GPUs on optimus systems
The d76 version of Mesa introduced a change which seems to select the Nouveau
driver where it would previously select the Intel driver on Optimus systems.
This is in conflict with Chrome which is hardcoded to use /dev/dri/card0 (Intel
on Optimus systems) which causes the graphics stack to hang.

OVER-10149