Skip the first radeonsi driver of a multi-gpu system.
Skip the first radeonsi driver similar to how it has been done for
nouveau systems. This avoids issues found on a dual GPU machine
which has an intel device as well as a radeon device.
The result of this was causing flickers with the i965 driver that
was being used.

Additionally clean up the fd and memory that is leaked when the gpu is
skipped.

OVER-11316

(cherry picked from commit afc1207f9399d53f9ac5c2b1dc0802c6fab5c6d8)